+++ packages/core/src/index.ts
@@ -1,4 +1,4 @@
-export const VERSION = "0.6.37";
+export const VERSION = "0.6.38";
 
 export type {
 	CreateComponentArgs,